+++ src/store/todos.js
@@ -22,14 +22,14 @@ const todos = {
         },
         setDraggedState({ commit }, [fromIndex, toIndex]) {
             const currentItem = this.getters.allTodos[fromIndex];
-            const currentTodos = this.getters.allTodos
-            currentTodos.splice(fromIndex, 1)
-			currentTodos.splice(toIndex, 0, currentItem)
+            const currentTodos = this.getters.allTodos;
+            currentTodos.splice(fromIndex, 1);
+            currentTodos.splice(toIndex, 0, currentItem);
             commit('updateTodos', currentTodos);
         },
-        addNewTaskAction({commit}, task){
-            commit("addNewTask", task)
-        }
+        addNewTaskAction({ commit }, task) {
+            commit('addNewTask', task);
+        },
     },
     mutations: {
         updateTodos(state, todos) {
@@ -79,17 +79,16 @@ const todos = {
         allTodosCounter(state) {
             return state.todos.length;
         },
-        filteredTodos: state => filter => {
+        filteredTodos: (state) => (filter) => {
             switch (filter) {
-                case "done":
-                        return state.todos.filter(item => item.isDone === true)
-                case "undone":
-                    return state.todos.filter(item => item.isDone === false)
+                case 'done':
+                    return state.todos.filter((item) => item.isDone === true);
+                case 'undone':
+                    return state.todos.filter((item) => item.isDone === false);
                 default:
-                    return state.todos
+                    return state.todos;
             }
-        }
+        },
     },
-}
-
+};
 export default todos;
